Add files via upload
This script uses a common threshold for the three planes to tag the ends of pixels corresponding to charge on the planes.  The 'continue' statements for the wire number correspond to the given sensitivities in the wire matching algo, and the 'continue' statements in the pixel matching for planes two and three correspond to the cutoffs beyond which pixels no longer correspond to wires for the file ccqe.root.