Add files via upload
Contains a function that matches wires on the first plane to others on the second and third planes with y and z coordinates within the given sensitivities.  The second function assigns the second and third plane wires to the first plane wire based on the information from a list resulting from the first function.